bug fixes to CSR

+++ erp_commerce/public/javascripts/extensions/compass_ae/erp_app/desktop/applications/product_manager/product_pricing_panel.js
@@ -147,7 +147,7 @@ Ext.define("Compass.ErpApp.Desktop.Applications.ProductManager.ProductPricingPan
         },
         fields:[{
           name:'price',
-          type:'decimal'
+          type:'float'
         }, 'currency', 'currency_display', 'from_date', 'thru_date', 'description','comments','pricing_plan_id']
       })
     });
